api/requirements.txt
django
djangorestframework
python-decouple
dj-database-url
dj_static
Pillow